YoutubeVideo-to-quiz-generator-main/requirements.txt
openai
streamlit
python-dotenv
youtube-transcript-api
langchain_community